django-backend/requirements.txt
asgiref
bcrypt
Django
django-cors-headers
django-postgresql
djangorestframework
passlib
psycopg2
PyJWT
python-dotenv
pytz
sqlparse
typing_extensions
tzdata
gunicorn==19.7.1
